requirements.txt
fastapi 
uvicorn 
crewai 
openai 
python-dotenv
langchain_community
duckduckgo-search
langchain
crewai-tools
langchain-ollama 
langchain-community 
crewai
torch
sse-starlette